Added 'Turn Over' overlay to repair skills
When Into the Breach is played in a very small window, the ui a unit changes. When a unit with a single weapon has spent its turn, the text "Turn Over" overlaps the repair skill.

Added an overlay to rebuild the missing text from displaying a custom repair icon over the vanilla repair icon.